Add checker and spaceshipRepo dependency in constructor

+++ src/main/java/dev/ime/solar_fleet/service/ShipClassServiceImpl.java
@@ -7,6 +7,8 @@ import org.bson.types.ObjectId;
 
 import dev.ime.solar_fleet.entity.ShipClass;
 import dev.ime.solar_fleet.repository.ShipClassRepository;
+import dev.ime.solar_fleet.repository.SpaceshipRepository;
+import dev.ime.solar_fleet.tool.Checker;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
 
@@ -16,18 +18,26 @@ public class ShipClassServiceImpl implements GenericService<ShipClass>{
 
 	
 	private final ShipClassRepository shipClassRepository;
+	private final SpaceshipRepository spaceshipRepository;
+	private final Checker checker;
 	
 	@Inject
-	public ShipClassServiceImpl(ShipClassRepository shipClassRepository) {
-		super();
+	public ShipClassServiceImpl(ShipClassRepository shipClassRepository, Checker checker,SpaceshipRepository spaceshipRepository) {
 		this.shipClassRepository = shipClassRepository;
+		this.checker = checker;
+		this.spaceshipRepository = spaceshipRepository;
 	}
 
 	@Override
 	public List<ShipClass> getAll() {
-		return shipClassRepository.findAll().list();
+		return shipClassRepository.listAll();
 	}
-
+	
+	@Override
+	public List<ShipClass> getAllPaged(int page) {
+		return shipClassRepository.findAll().page(page - 1, 5).list();
+	}
+	
 	@Override
 	public Optional<ShipClass> getById(ObjectId id) {
 		return Optional.ofNullable(shipClassRepository.findById(id));
@@ -55,8 +65,18 @@ public class ShipClassServiceImpl implements GenericService<ShipClass>{
 	}
 
 	@Override
-	public int delete(ObjectId id) {
-		return shipClassRepository.deleteById(id)?0:1;
+	public int delete(ObjectId id) {		
+		
+		if ( !checker.checkObjectId(id) ) {
+	        return 1;
+	    }
+		
+		Optional<ShipClass> opt = Optional.ofNullable(shipClassRepository.findById(id));
+		if ( opt.isPresent() && spaceshipRepository.list("shipClassId", id).isEmpty() ) {
+	        return shipClassRepository.deleteById(id) ? 0 : 1;
+	    }
+		
+		return 1;
 	}
 
 }
